DirCacheTree: fix possible AIOOBE in validate()
For index files which contain files and sub-directories with the same
name DirCacheTree.validate() fails with an ArrayIndexOutOfBoundsException.
By re-arranging checks we can avoid this failing comparison. This makes
the logic also more similar to CGit which will also check first for
files directly contained in the current cache tree and in this case
immediately advance to the next file without even looking at possible
equally named sub trees.

Bug: 552214
Change-Id: I6123a17f49379fd99806dd72406e30f308c79467
Signed-off-by: Marc Strapetz <marc.strapetz@syntevo.com>