DirCacheTree: better compatibility with CGit
CGit distinguishes between dircache states for which there exists a
valid cache tree and for those where no cache tree exists (this is basically
the case for unmerged paths). It will only update the dircache tree
for valid states [1]. It could possibly still read and write arbitrary
cache trees (even those which can't be updated) but this won't do any
harm because all code paths which are actually relying on valid cache
tree data (especially OIDs) will make sure to update the cache tree
before.

JGit is using cache trees partly as CGit does but also for its internal
processing, especially in DirCacheIterator. DirCacheIterator relies
on cache trees which are updated even for cache states which are invalid
from CGit's perspective, e.g. in case of conflicts. The actual
validation only occurs when writing the cache tree to the ObjectInserter
using DirCache.writeTree (note that DirCache.validate() is not doing
this actual validation).

Thus, JGit ensures to only write valid cache trees to the object database,
but the cache tree instance in DirCache may already be in an invalid state
and could be written out to disk by DirCache.writeTo.
It's unclear whether the current API allows to write valid cache trees
(with OID and entrySpan >= 0) for invalid cache states.
What is definitely possible is that obsolete invalid cache trees
(no OID, entrySpan = -1) are written to disk, as
DirCacheTreeTest.testDirCacheTreeNotTouchedByIterator shows.

To clarify the different usages of cache trees within JGit and prevent
writing of invalid cache trees to disk, the DirCache API for cache trees
is now split up.

[1] cache_tree_update in cache-tree.c (as of commit cd9994f9)

Change-Id: Ie0cbe71c002bc49ca79a9c13a843eee1fb3c7a11
Signed-off-by: Marc Strapetz <marc.strapetz@syntevo.com>